+++ wcag-tool/chrome-extension/src/components/editor/LabelList.jsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import LabelListItem from './LabelListItem'
+
+export default function LabelList({ labels }) {
+  return labels.map((label) => <LabelListItem key={label} label={label} />)
+}
